Fix pointer truncation for x64 build (fixes #11)
The x86 code often stored/cast host pointers into 32-bit types
(uint32_t/DWORD/LONG/int), which breaks on x64 where pointers are
64-bit. Fix all such places in code compiled for the x64 build:

- cart: keep the CreateFile/CreateFileMapping HANDLEs as HANDLE
  instead of uint32_t (they were truncated, breaking cart mapping
  on x64). cart.h now includes <windows.h> for the HANDLE type.
- ultrahle: Controller1Prop took LPARAM pointers (PROPSHEETPAGE,
  NMHDR) in a 32-bit LONG, truncating them on x64. Use LPARAM and
  make the dialog procs return LRESULT like DLGPROC expects.
- cmd/disasm: the JIT debug disassembly printed and passed compiled
  code addresses through int/uint32_t. Use intptr_t for the code
  offset and %p for host pointers.
- XOpenGL/fxgeom: keep the decompiled setuprvx semantics but route
  the pointer round-trips through uintptr_t/intptr_t.

The dynamic recompiler (cpua.c/cpuanew.c/cpuautil.c) still uses
32-bit pointer tricks in the x86 code it generates; those files stay
excluded from the x64 build, which runs the portable C interpreter
(cpuc.c) instead.

+++ src/disasm.h
@@ -18,7 +18,7 @@ char* disasm(uint32_t pos, uint32_t x);
 char* disasmrsp(uint32_t pos, uint32_t x);
 
 // unassemble X86 code
-char* disasmx86(uint8_t* opcode, int codeoff, int* len);
+char* disasmx86(uint8_t* opcode, intptr_t codeoff, int* len);
 
 // dump memory range into a file (two ranges; one for code, one for data)
 void  disasm_dumpcode(char* filename, uint32_t addr, int size, uint32_t dataaddr, int datasize);
